--- a/package.json
+++ b/package.json
@@ -130,7 +130,7 @@
     "start": "export NODE_ENV=development; check-engines && check-dependencies && webpack-dashboard -p 3736 -- webpack-dev-server --config ./webpack.dev.js",
     "_build": "export HEAD_COMMIT=$(git rev-parse --short HEAD); check-engines && check-dependencies && rimraf dist; webpack --config ./webpack.build.js --progress --profile --colors",
     "serve-static": "export NODE_ENV=staging; check-engines && check-dependencies && npm run _build && node ./static-server.js",
-    "_publish": "s3-batch-upload -b zooniverse-static -p ./dist -C 200 -acl 'public-read' -r \"$PATH_ROOT/$SUBDIR\"",
+    "_publish": "s3-batch-upload -b zooniverse-static -p ./dist -C 200 -a '{\"**/index.html\": \"no-cache, must-revalidate\", \"**/*.*\": \"max-age=86400\"}' -acl 'public-read' -r \"$PATH_ROOT/$SUBDIR\"",
     "_log-staging-url": "echo \"Staged at https://$SUBDIR.pfe-preview.zooniverse.org/\"",
     "_stage": "export PATH_ROOT=preview.zooniverse.org/panoptes-front-end; npm run _build && npm run _publish && npm run _log-staging-url",
     "stage-with-jenkins": "export NODE_ENV=staging; export SUBDIR=$(echo $BRANCH_NAME | awk '{print tolower($0)}'); npm run _stage",