@srallen 2020-11-12
Could we check in with the migration status?  If we're close to moving static assets over to Azure, then it make sense to switch this to an Azure CLI at that time than an updated S3 one.