@eatyourgreens 2020-11-12
[s3-batch-upload](https://github.com/mediamonks/s3-batch-upload) has a `--dry-run` option, which might be useful for testing this without actually overwriting the live site.