@eatyourgreens 2020-11-12
@srallen the last conversation I had with Adam, we talked about making sure that the static sites all had Docker builds and Jenkinsfiles. I think we're on S3 for the near future but, yeah, it's worth thinking about how we'll copy files to Azure.

`publisssh` is very much out-of-date, so I'd consider upgrading at the same time as adding Jenkins builds to the static repos. I've just finished doing that for Pandora.