@srallen 2020-11-12
Ok, we should check in with the ops team. Presumably that information is at least several weeks out of date. @zwolf can you provide us an update with the migration of static assets to Azure?