@eatyourgreens 2020-11-12
Here’s the ops issue for switching over. It’s open for discussion at the moment. https://github.com/zooniverse/operations/issues/494