@eatyourgreens 2020-11-12
I’ve posted a couple of useful links, over on that issue, for Azure Storage node clients.