@eatyourgreens 2020-11-16
lab.zooniverse.org is deploying automatically with s3-batch-upload (https://github.com/zooniverse/pfe-lab/pull/214) and translations.zooniverse.org too (https://github.com/zooniverse/pandora/pull/166.) This branch deployed to https://pr-5840.pfe-preview.zooniverse.org without problems, so I think we are good to go.