@eatyourgreens 2020-11-16
Let me double-check the deployed commit, because this is what I'm seeing on in the HTML on the staging branch.
```html
<!--Last build: Fri Nov 13 2020 11:22:16 GMT+0000 (Coordinated Universal Time)-->
<!--Deployed commit: 1d59a980-->
<!--Build env: staging-->
```