@eatyourgreens 2020-11-16
A production deploy with the `--dry-run` flag runs without errors.
```
asosxlap127:Panoptes-Front-End odonnell$ npm run deploy

> panoptes-front-end@ deploy /Users/odonnell/zooniverse/Panoptes-Front-End
> export NODE_ENV=production; npm run _deploy


> panoptes-front-end@ _deploy /Users/odonnell/zooniverse/Panoptes-Front-End
> export PATH_ROOT=www.zooniverse.org; npm run _build && npm run _publish


> panoptes-front-end@ _build /Users/odonnell/zooniverse/Panoptes-Front-End
> export HEAD_COMMIT=$(git rev-parse --short HEAD); check-engines && check-dependencies && rimraf dist; webpack --config ./webpack.build.js --progress --profile --colors

49758ms building                                                                    
16ms finish module graph                            
12ms sealing                               
0ms basic dependencies optimization 
22ms dependencies optimization                          
0ms advanced dependencies optimization 
0ms after dependencies optimization 
22ms chunk graph
1ms after chunk graph                          
1ms optimizing 
0ms basic module optimization 
0ms module optimization 
0ms advanced module optimization 
0ms after module optimization 
1ms basic chunk optimization                             
0ms chunk optimization 
37ms advanced chunk optimization                        
0ms after chunk optimization 
0ms module and chunk tree optimization 
0ms after module and chunk tree optimization 
0ms basic chunk modules optimization 
53ms chunk modules optimization                          
1ms advanced chunk modules optimization 
0ms after chunk modules optimization 
7ms module reviving                 
13ms module order optimization                               
1ms advanced module order optimization 
0ms before module ids 
0ms module ids 
7ms module id optimization 
1ms chunk reviving                 
0ms chunk order optimization                               
0ms before chunk ids 
3ms chunk id optimization                          
6ms after chunk id optimization 
3ms record modules                 
0ms record chunks                 
46ms hashing
11ms content hashing                        
0ms after hashing 
0ms record hash 
1ms module assets processing 
264ms chunk assets processing
7ms additional chunk assets processing 
0ms recording 
1ms additional asset processing 
23234ms chunk asset optimization          
0ms after chunk asset optimization 
0ms asset optimization 
0ms after asset optimization 
1ms after seal 
873ms emitting                 
13ms after emitting                 
Hash: dcda24474baf98338525
Version: webpack 4.29.6
Time: 74449ms
Built at: 16/11/2020 17:59:55
                                                                     Asset       Size  Chunks                    Chunk Names
                                                     assets/about-hero.png    362 KiB          [emitted]  [big]  
                                              assets/all-projects-hero.png    171 KiB          [emitted]         
                                                assets/builder-landing.jpg    297 KiB          [emitted]  [big]  
                                         assets/default-project-avatar.png  949 bytes          [emitted]         
                                     assets/default-project-background.jpg    107 KiB          [emitted]         
                       assets/dev-classifier/1715_labs/building_plans.jpeg   96.7 KiB          [emitted]         
            assets/dev-classifier/1715_labs/greenwich_satellite_image.jpeg    295 KiB          [emitted]  [big]  
                   assets/dev-classifier/1715_labs/london_street_view.jpeg    301 KiB          [emitted]  [big]  
                                        assets/dev-classifier/algernon.txt  942 bytes          [emitted]         
                                      assets/dev-classifier/landscape.jpeg   66.3 KiB          [emitted]         
                                       assets/dev-classifier/portrait.jpeg   79.8 KiB          [emitted]         
                                          assets/dev-classifier/small.jpeg     28 KiB          [emitted]         
                                       assets/dev-classifier/testData.json   4.34 KiB          [emitted]         
                                      assets/dev-classifier/testData0.json   4.34 KiB          [emitted]         
                                      assets/dev-classifier/very-tall.jpeg    189 KiB          [emitted]         
                                      assets/dev-classifier/very-wide.jpeg    218 KiB          [emitted]         
                                                  assets/donate-button.png   21.3 KiB          [emitted]         
                                         assets/featured-projects/1104.jpg   72.7 KiB          [emitted]         
                                           assets/featured-projects/21.jpg   59.5 KiB          [emitted]         
                                         assets/featured-projects/3063.jpg   72.1 KiB          [emitted]         
                                         assets/featured-projects/3098.jpg   98.9 KiB          [emitted]         
                                          assets/featured-projects/376.jpg   46.6 KiB          [emitted]         
                                         assets/featured-projects/4973.jpg   56.4 KiB          [emitted]         
                                           assets/featured-projects/60.jpg     83 KiB          [emitted]         
                                         assets/featured-projects/6263.jpg     26 KiB          [emitted]         
    assets/featured-projects/featured-project-20170328-stargazing-live.jpg    278 KiB          [emitted]  [big]  
    assets/featured-projects/featured-project-20170403-stargazing-live.jpg    420 KiB          [emitted]  [big]  
   assets/featured-projects/featured-project-20170531-galaxy-nurseries.jpg   89.1 KiB          [emitted]         
          assets/featured-projects/featured-project-20170711-milky-way.jpg   86.6 KiB          [emitted]         
            assets/featured-projects/featured-project-20170801-steller.jpg    238 KiB          [emitted]         
assets/featured-projects/featured-project-20170829-steelpan_vibrations.jpg   50.3 KiB          [emitted]         
               assets/featured-projects/featured-project-20170913-irma.jpg   27.8 KiB          [emitted]         
        assets/featured-projects/featured-project-20171023-autumnwatch.jpg     74 KiB          [emitted]         
 assets/featured-projects/featured-project-20171204-fossil_atmospheres.jpg   63.8 KiB          [emitted]         
    assets/featured-projects/featured-project-20180108-project_plumage.jpg    176 KiB          [emitted]         
       assets/featured-projects/featured-project-20180223-scotus-notes.jpg   50.7 KiB          [emitted]         
                                               assets/feedback-details.png   26.8 KiB          [emitted]         
                                               assets/feedback-summary.png    100 KiB          [emitted]         
                                                assets/feedback-toggle.png   49.6 KiB          [emitted]         
                                           assets/highlights-book-2019.png    128 KiB          [emitted]         
                                           assets/highlights-book-2020.jpg   1.32 MiB          [emitted]  [big]  
                                                  assets/home-appStore.png   13.7 KiB          [emitted]         
                                               assets/home-collaborate.gif    351 KiB          [emitted]  [big]  
                                                  assets/home-computer.png    780 KiB          [emitted]  [big]  
                                                assets/home-contribute.gif    558 KiB          [emitted]  [big]  
                                                  assets/home-discover.gif   67.1 KiB          [emitted]         
                                                   assets/home-explore.gif    217 KiB          [emitted]         
                                                assets/home-googlePlay.png   17.5 KiB          [emitted]         
                                              assets/home-researchers1.jpg    612 KiB          [emitted]  [big]  
                                              assets/home-researchers2.jpg    660 KiB          [emitted]  [big]  
                                              assets/home-researchers3.jpg    563 KiB          [emitted]  [big]  
                                                     assets/home-video.jpg   87.8 KiB          [emitted]         
                                                     assets/home-video.mp4   1.86 MiB          [emitted]  [big]  
                                                     assets/home-video.ogv   7.16 MiB          [emitted]  [big]  
                                                    assets/home-video.webm    5.2 MiB          [emitted]  [big]  
                                            assets/how-to-lab/how-to-1.jpg    107 KiB          [emitted]         
                                           assets/how-to-lab/how-to-10.jpg    186 KiB          [emitted]         
                                           assets/how-to-lab/how-to-11.jpg   26.3 KiB          [emitted]         
                                           assets/how-to-lab/how-to-12.jpg   85.5 KiB          [emitted]         
                                           assets/how-to-lab/how-to-13.jpg    121 KiB          [emitted]         
                                            assets/how-to-lab/how-to-2.jpg   74.6 KiB          [emitted]         
                                            assets/how-to-lab/how-to-3.jpg    154 KiB          [emitted]         
                                            assets/how-to-lab/how-to-4.jpg    167 KiB          [emitted]         
                                            assets/how-to-lab/how-to-5.jpg    135 KiB          [emitted]         
                                            assets/how-to-lab/how-to-6.jpg    173 KiB          [emitted]         
                                            assets/how-to-lab/how-to-7.jpg    202 KiB          [emitted]         
                                            assets/how-to-lab/how-to-8.jpg    127 KiB          [emitted]         
                                            assets/how-to-lab/how-to-9.jpg    191 KiB          [emitted]         
                                                  assets/mobile-sample.png   43.1 KiB          [emitted]         
                                                  assets/simple-avatar.jpg   5.69 KiB          [emitted]         
                                                  assets/simple-avatar.png   11.7 KiB          [emitted]         
                                                 assets/simple-pattern.jpg   7.47 KiB          [emitted]         
                                                 assets/simple-pattern.png   3.59 KiB          [emitted]         
                                                      assets/team/adam.jpg   2.23 KiB          [emitted]         
                                                      assets/team/alex.jpg   11.4 KiB          [emitted]         
                                               assets/team/alex_bowyer.jpg   2.01 KiB          [emitted]         
                                                       assets/team/ali.jpg   3.01 KiB          [emitted]         
                                                      assets/team/amit.jpg   1.61 KiB          [emitted]         
                                                       assets/team/amy.png   8.96 KiB          [emitted]         
                                                    assets/team/andrea.jpg   24.8 KiB          [emitted]         
                                                     assets/team/arfon.jpg   6.78 KiB          [emitted]         
                                                     assets/team/becky.jpg   6.49 KiB          [emitted]         
                                               assets/team/beckyrother.jpg     89 KiB          [emitted]         
                                                  assets/team/beverley.jpg   21.1 KiB          [emitted]         
                                                     assets/team/brian.jpg   9.07 KiB          [emitted]         
                                                    assets/team/brooke.jpg   7.08 KiB          [emitted]         
                                                       assets/team/cam.jpg    6.1 KiB          [emitted]         
                                                   assets/team/chris-d.jpg   20.7 KiB          [emitted]         
                                                   assets/team/chris-l.png   13.9 KiB          [emitted]         
                                                    assets/team/chriss.png     13 KiB          [emitted]         
                                                   assets/team/coleman.jpg   3.62 KiB          [emitted]         
                                                    assets/team/darren.jpg   35.3 KiB          [emitted]         
                                                        assets/team/ed.png   18.3 KiB          [emitted]         
                                                     assets/team/grant.jpg   4.96 KiB          [emitted]         
                                                      assets/team/greg.jpg    2.5 KiB          [emitted]         
                                                  assets/team/hannah_s.jpg   11.6 KiB          [emitted]         
                                                     assets/team/heath.png   17.6 KiB          [emitted]         
                                               assets/team/helenspiers.jpg   30.5 KiB          [emitted]         
                                                      assets/team/hugh.jpg   5.17 KiB          [emitted]         
                                                     assets/team/james.jpg   36.4 KiB          [emitted]         
                                                 assets/team/jen_gupta.jpg   7.21 KiB          [emitted]         
                                                       assets/team/jim.jpg   6.31 KiB          [emitted]         
                                                    assets/team/jordan.jpg   21.3 KiB          [emitted]         
                                                     assets/team/julie.png   15.5 KiB          [emitted]         
                                                     assets/team/karen.jpg   11.1 KiB          [emitted]         
                                                     assets/team/kelly.jpg   7.28 KiB          [emitted]         
                                                      assets/team/kyle.jpg   10.1 KiB          [emitted]         
                                                     assets/team/laura.jpg   13.9 KiB          [emitted]         
                                                   assets/team/laura_w.jpg   9.53 KiB          [emitted]         
                                                      assets/team/lucy.jpg   10.2 KiB          [emitted]         
                                                  assets/team/mariam_s.jpg    2.8 KiB          [emitted]         
                                                      assets/team/mark.jpg   20.2 KiB          [emitted]         
                                                    assets/team/marten.jpg   5.13 KiB          [emitted]         
                                                       assets/team/meg.jpg   20.9 KiB          [emitted]         
                                                   assets/team/melanie.jpg   2.09 KiB          [emitted]         
                                                   assets/team/michael.png   3.26 KiB          [emitted]         
                                                      assets/team/noah.jpg   12.8 KiB          [emitted]         
                                                     assets/team/perry.jpg   10.3 KiB          [emitted]         
                                                       assets/team/rob.jpg   1.41 KiB          [emitted]         
                                                     assets/team/robin.jpg   12.4 KiB          [emitted]         
                                                     assets/team/roger.jpg   24.6 KiB          [emitted]         
                                                      assets/team/rona.jpg   7.63 KiB          [emitted]         
                                                       assets/team/sam.jpg   91.5 KiB          [emitted]         
                                                    assets/team/samuel.jpg   81.5 KiB          [emitted]         
                                                    assets/team/sandor.jpg   1.18 KiB          [emitted]         
                                                     assets/team/sarah.jpg   13.3 KiB          [emitted]         
                                                    assets/team/sascha.png   17.3 KiB          [emitted]         
                                             assets/team/shaunanoordin.jpg     52 KiB          [emitted]         
                                                    assets/team/simone.jpg     17 KiB          [emitted]         
                                                     assets/team/steve.png   16.1 KiB          [emitted]         
                                                    assets/team/stuart.png   14.5 KiB          [emitted]         
                                        assets/team/team-member-avatar.jpg   5.37 KiB          [emitted]         
                                                  assets/team/veronica.jpg   19.8 KiB          [emitted]         
                                                  assets/team/victoria.JPG   6.96 KiB          [emitted]         
                                                      assets/team/will.jpg   31.1 KiB          [emitted]         
                                                      assets/team/zach.JPG   9.02 KiB          [emitted]         
                                            assets/user-header-default.jpg    149 KiB          [emitted]         
                                                   assets/zoohome-hero.jpg   63.7 KiB          [emitted]         
                                                               favicon.ico    5.3 KiB          [emitted]         
                                                                index.html   3.98 KiB          [emitted]         
                                             main-9cda41f726b4accd5caa.css    288 KiB       0  [emitted]  [big]  main
                                              main-d73f7f3f159ca9c10627.js   3.44 MiB       0  [emitted]  [big]  main
                                           vendor-8786d83f76a19bbbc727.css   25.5 KiB       1  [emitted]         vendor
                                            vendor-f74b8e1c4a8f959319d2.js   1.99 MiB       1  [emitted]  [big]  vendor
Entrypoint main [big] = vendor-8786d83f76a19bbbc727.css vendor-f74b8e1c4a8f959319d2.js main-9cda41f726b4accd5caa.css main-d73f7f3f159ca9c10627.js
   [6] ./node_modules/react-router/es/index.js + 32 modules 76.4 KiB {1} [built]
       |    33 modules
  [18] ./node_modules/react-redux/es/index.js + 19 modules 37 KiB {1} [built]
       |    20 modules
  [64] ./app/redux/ducks/classify.js 22.3 KiB {0} [built]
       [810] 2ms -> [811] 6022ms -> [174] 17064ms -> factory:4790ms building:0ms dependencies:33729ms = 61607ms
…